refactor(dry): eliminate cross-repo and internal code duplication
- Remove dead core.py (192 lines) shadowed by core/ package since Phase 2 split
- Replace 3 duplicated get_logger try/except fallback patterns in model_explorer
  with centralized logger_utils (which already has built-in fallback support)
- Replace inline logging.basicConfig calls in tool entry points
  (video_analyzer, data_explorer, model_generation CLI) with shared logging_config
- Create upstream-drift-shared pyproject.toml for cross-repo pip consumption (DRY-1)
- Document source-of-truth ownership in 8 shared module __init__.py files (DRY-2)

Closes #1266
