+++ src/shared/python/engine_core/__init__.py
@@ -1,3 +1,7 @@
-"""Engine interfaces, registry, manager, and loaders."""
+"""Engine interfaces, registry, manager, and loaders.
+
+Source of Truth: UpstreamDrift (this repository)
+Cross-repo install: pip install upstream-drift-shared
+"""
 
 __all__: list[str] = []
