+++ src/shared/python/plotting/__init__.py
@@ -1,5 +1,11 @@
 """Plotting package for golf swing visualization.
 
+Source of Truth: UpstreamDrift (this repository)
+Consumers:
+    - Tools repo: src/shared/python/plot_theme/ (vendored copy)
+    - Gasification_Model: integrated via vendor/ud-tools/
+Cross-repo install: pip install upstream-drift-shared[plotting]
+
 This package provides modular plotting components:
 
 - base: MplCanvas, RecorderInterface, color schemes
