+++ src/shared/python/security/__init__.py
@@ -1,3 +1,8 @@
-"""Security utilities, secure subprocess, and environment validation."""
+"""Security utilities, secure subprocess, and environment validation.
+
+Source of Truth: UpstreamDrift (this repository)
+Consumers: Tools (security module), Gasification_Model (security utilities)
+Cross-repo install: pip install upstream-drift-shared
+"""
 
 __all__: list[str] = []
