+++ src/shared/python/validation_pkg/__init__.py
@@ -1,3 +1,8 @@
-"""Validation, statistical analysis, and data fitting."""
+"""Validation, statistical analysis, and data fitting.
+
+Source of Truth: UpstreamDrift (this repository)
+Consumers: Tools, Gasification_Model (via upstream-drift-shared package)
+Cross-repo install: pip install upstream-drift-shared
+"""
 
 __all__: list[str] = []
